% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{aggregate_networks}
\alias{aggregate_networks}
\title{Aggregate networks}
\usage{
aggregate_networks(adj_list, method = "ss_debias", verbose = F)
}
\arguments{
\item{adj_list}{list of adjacency matrices}

\item{method}{method, either \code{"ss_debias"}, \code{"ss"} or \code{"sum"}}

\item{verbose}{boolean}
}
\value{
a matrix
}
\description{
Aggregate networks
}
