% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{compute_markov_transition}
\alias{compute_markov_transition}
\title{Compute the appropriate transition matrix to form desired stationary vector}
\usage{
compute_markov_transition(vec, delta, sig_dig = 3)
}
\arguments{
\item{vec}{desired stationary vector of length 3}

\item{delta}{number between 0 and 1 (inclusive)}

\item{sig_dig}{number of significant digits}
}
\value{
a row-stochastic matrix
}
\description{
Authors note: This does NOT always work. To fix this, I suspect
I would need to reparameterize the problem such that the D matrix
is positive definite (by having only 3 unknowns instead of 6).
}
