% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_methods.R
\name{coreg}
\alias{coreg}
\title{Co-regularized  spectral  clustering}
\usage{
coreg(x, k, beta, max_iter = 100, verbose = F)
}
\arguments{
\item{x}{list of adjacency matrices}

\item{k}{number of clusters}

\item{beta}{positive numeric}

\item{max_iter}{positive integer}

\item{verbose}{boolean}
}
\value{
list
}
\description{
The main Coreg function. Inputs are the list of adjacency matrices x, number of nodes n, nunber of communities
k and regularization parameter beta.
}
