% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{.svd_projection}
\alias{.svd_projection}
\title{Do an SVD projection}
\usage{
.svd_projection(mat, K, weighted = F)
}
\arguments{
\item{K}{positive integer less than \code{n}}

\item{weighted}{boolean}

\item{adj_mat}{numeric matrix with \code{n} rows and \code{n} columns}
}
\value{
numeric matrix
}
\description{
Uses \code{RSpectra::svds} to compute the \code{k} leading singular vectors, but
sometimes there are numerical instability issues. In case of crashes, the code
then uses the default \code{svd} function.
}
