% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{flatten}
\alias{flatten}
\title{Flatten collection of adjacency matrices}
\usage{
flatten(adj_list)
}
\arguments{
\item{adj_list}{list of adjacency matrices}
}
\value{
a matrix
}
\description{
Flatten collection of adjacency matrices
}
