% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_clustering.R
\name{greedy_clustering}
\alias{greedy_clustering}
\title{Greedy clustering}
\usage{
greedy_clustering(adj_list, K, nstart = 10, iter_max = 100, verbose = F)
}
\arguments{
\item{adj_list}{list of adjacency matrices}

\item{K}{positive integer}

\item{nstart}{positive integer}

\item{iter_max}{positive integer}

\item{verbose}{boolean}
}
\value{
membership vector
}
\description{
Greedy clustering
}
