% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_methods.R
\name{lmfo}
\alias{lmfo}
\title{Linked matrix factorization}
\usage{
lmfo(x, k)
}
\arguments{
\item{x}{list of adjacency matrices}

\item{k}{number of clusters}
}
\value{
list
}
\description{
The main function with BFGS optimization
}
