% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{spectral_clustering}
\alias{spectral_clustering}
\title{Spectral clustering}
\usage{
spectral_clustering(sym_mat, K, weighted = F)
}
\arguments{
\item{sym_mat}{symmetric matrix}

\item{K}{positive integer}

\item{weighted}{boolean}
}
\value{
membership vector
}
\description{
Spectral clustering
}
